+++ src/main/java/tfar/classicbar/compat/Helpers.java
@@ -1,7 +1,6 @@
 package tfar.classicbar.compat;
 
 import de.teamlapen.vampirism.api.VReference;
-import de.teamlapen.vampirism.api.VampirismAPI;
 import de.teamlapen.vampirism.api.entity.factions.IFactionPlayerHandler;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.util.Direction;
@@ -13,6 +12,8 @@ import net.minecraftforge.fml.ModList;
 public class Helpers {
 
 	public static final boolean vampirismloaded = ModList.get().isLoaded("vampirism");
+	public static final boolean elenaiDodgeLoaded = ModList.get().isLoaded("elenaidodge2");
+
 
 	@CapabilityInject(IFactionPlayerHandler.class)
 	private static Capability<IFactionPlayerHandler> CAP_FACTION_HANDLER_PLAYER = null;
